backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
google-genai==1.14.0
python-dotenv==1.0.1
numpy==2.2.1
pydantic==2.10.4
httpx==0.28.1
lru-dict==1.3.0
